Changed the file administration, fixed the new imports, removed unused code, added the start of the picklist component
