+++ src/components/PickList/PickList.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PickList = (props) => {
+  return <div>PickList</div>;
+};
+
+export default PickList;
